+++ inst/temp/UtilFomErrorClassFile.R
@@ -1,16 +1,14 @@
 library(RJafroc)
 rm(list = ls())
+
 fileName <- system.file("extdata", "/toyFiles/FROC/frocLocatClass.xlsx", package = "RJafroc", mustWork = TRUE)
-x <- DfReadDataFile(fileName = fileName, newExcelFileFormat = T)
+#fileName <- system.file("extdata", "/toyFiles/FROC/frocCr.xlsx", package = "RJafroc", mustWork = TRUE)
 
-print(UtilFigureOfMerit(x, FOM = "wAFROC1"))
+x1 <- DfReadDataFile(fileName = fileName, newExcelFileFormat = T)
+print(UtilFigureOfMerit(x1, FOM = "wAFROC1"))
 
+x2 <- DfReadDataFile(fileName = fileName, newExcelFileFormat = F)
+print(UtilFigureOfMerit(x2, FOM = "wAFROC1"))
 
-# fileName <- system.file(
-#   "extdata", "/toyFiles/FROC/frocCr.xlsx", package = "RJafroc", mustWork = TRUE)
-# 
-# x <- DfReadDataFile(fileName = fileName)
-# 
-# UtilFigureOfMerit(x, FOM = "wAFROC1")